# [gpio] Refactor GPIO and Signal implementation