--- a/src/modm/board/devebox_stm32f4xx/board.hpp
+++ b/src/modm/board/devebox_stm32f4xx/board.hpp
@@ -137,9 +137,6 @@ initialize()
 	LedGreen::setOutput(modm::Gpio::Low);
 
 	Button::setInput(Gpio::InputType::PullDown);
-	Button::setInputTrigger(Gpio::InputTrigger::RisingEdge);
-	Button::enableExternalInterrupt();
-//	Button::enableExternalInterruptVector(12);
 }