--- a/src/modm/board/nucleo_f429zi/board.hpp
+++ b/src/modm/board/nucleo_f429zi/board.hpp
@@ -147,9 +147,6 @@ initialize()
     LedRed::setOutput(modm::Gpio::Low);
 
     Button::setInput();
-    Button::setInputTrigger(Gpio::InputTrigger::RisingEdge);
-    Button::enableExternalInterrupt();
-//  Button::enableExternalInterruptVector(12);
 }
 
 inline void