--- a/src/modm/board/nucleo_f746zg/board.hpp
+++ b/src/modm/board/nucleo_f746zg/board.hpp
@@ -142,9 +142,6 @@ initialize()
     LedRed::setOutput(modm::Gpio::Low);
 
     Button::setInput();
-    Button::setInputTrigger(Gpio::InputTrigger::RisingEdge);
-    Button::enableExternalInterrupt();
-//  Button::enableExternalInterruptVector(12);
 }
 
 }