--- a/src/modm/board/nucleo_h743zi/board.hpp
+++ b/src/modm/board/nucleo_h743zi/board.hpp
@@ -185,9 +185,6 @@ initialize()
     LedRed::setOutput(modm::Gpio::Low);
 
     Button::setInput();
-    Button::setInputTrigger(Gpio::InputTrigger::RisingEdge);
-    Button::enableExternalInterrupt();
-//  Button::enableExternalInterruptVector(12);
 }
 
 /// FIXME: USB does not work on this board.