--- a/src/modm/board/nucleo_l496zg-p/board.hpp
+++ b/src/modm/board/nucleo_l496zg-p/board.hpp
@@ -162,9 +162,6 @@ initialize()
 	LedRed::setOutput(modm::Gpio::Low);
 
 	Button::setInput();
-//	Button::setInputTrigger(Gpio::InputTrigger::RisingEdge);
-//	Button::enableExternalInterrupt();
-//	Button::enableExternalInterruptVector(12);
 }
 
 inline void