--- a/src/modm/platform/adc/at90_tiny_mega/adc_mega.hpp.in
+++ b/src/modm/platform/adc/at90_tiny_mega/adc_mega.hpp.in
@@ -112,7 +112,7 @@ private:
 
 public:
 	// start inherited documentation
-	template< template<Peripheral _> class... Signals >
+	template< class... Signals >
 	static void
 	connect()
 	{