--- a/src/modm/platform/adc/at90_tiny_mega/adc_tiny.hpp.in
+++ b/src/modm/platform/adc/at90_tiny_mega/adc_tiny.hpp.in
@@ -158,7 +158,7 @@ private:
 
 public:
 	// start inherited documentation
-	template< template<Peripheral _> class... Signals >
+	template< class... Signals >
 	static void
 	connect()
 	{