--- a/src/modm/platform/comp/stm32/comp.hpp.in
+++ b/src/modm/platform/comp/stm32/comp.hpp.in
@@ -508,7 +508,7 @@ namespace modm::platform
 
 	public:
 		// start inherited documentation
-		template< template<Peripheral _> class... Signals >
+		template< class... Signals >
 		static void
 		connect()
 		{