--- a/src/modm/platform/dac/stm32/dac.hpp.in
+++ b/src/modm/platform/dac/stm32/dac.hpp.in
@@ -58,7 +58,7 @@ public:
 	};
 %% endif
 
-	template< template<Peripheral _> class... Signals >
+	template< class... Signals >
 	static void
 	connect()
 	{