--- a/src/modm/platform/dac/stm32/dac_dma.hpp.in
+++ b/src/modm/platform/dac/stm32/dac_dma.hpp.in
@@ -31,7 +31,7 @@ namespace modm::platform
 class Dac{{ id }}Dma
 {
 public:
-	template< template<Peripheral _> class... Signals >
+	template< class... Signals >
 	static void
 	connect()
 	{