--- a/src/modm/platform/extint/sam/module.lb
+++ b/src/modm/platform/extint/sam/module.lb
@@ -10,8 +10,6 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 # -----------------------------------------------------------------------------
 
-import sys
-
 def init(module):
     module.name = ":platform:extint"
     module.description = "External Interrupt"