--- a/src/modm/platform/gpio/at90_tiny_mega/base.hpp.in
+++ b/src/modm/platform/gpio/at90_tiny_mega/base.hpp.in
@@ -9,16 +9,12 @@
  */
 // ----------------------------------------------------------------------------
 
-#ifndef MODM_AVR_GPIO_BASE_HPP
-#define MODM_AVR_GPIO_BASE_HPP
+#pragma once
 
 #include "define.h"
 #include <modm/architecture/interface/gpio.hpp>
 
-namespace modm
-{
-
-namespace platform
+namespace modm::platform
 {
 
 /// @ingroup modm_platform_gpio
@@ -75,14 +71,6 @@ struct Gpio
 		{{ signal }},
 %% endfor
 	};
-
-protected:
-	/// @cond
-	static constexpr uint8_t
-	i(InputType config) { return uint8_t(config); }
-	static constexpr uint8_t
-	i(InputTrigger trigger) { return uint8_t(trigger); }
-	/// @endcond
 };
 
 /**
@@ -95,12 +83,12 @@ protected:
 template< class Pin >
 class GpioOpenDrain : public Pin
 {
-	static Gpio::InputType inputType;
+	static inline Gpio::InputType inputType = Gpio::InputType::Floating;
 	static_assert(Pin::direction == modm::Gpio::Direction::InOut, "Pin must inherit from modm::GpioIO");
 public:
-	using Output = GpioOpenDrain<typename Pin::IO>;
-	using Input = GpioOpenDrain<typename Pin::IO>;
 	using IO = GpioOpenDrain<typename Pin::IO>;
+	using Output = IO;
+	using Input = IO;
 	using Type = typename Pin::Type;
 
 	static constexpr modm::Gpio::Direction direction = modm::Gpio::Direction::Out;
@@ -113,42 +101,24 @@ public:
 	};
 
 public:
-	inline static void configure(Gpio::InputType type)
-	{ inputType = type; }
-	modm_always_inline static void setInput() {}
-	inline static void setInput(Gpio::InputType type)
-	{ inputType = type; }
-	modm_always_inline static void setOutput() {}
-	modm_always_inline static void setOutput(OutputType) {}
-	modm_always_inline static void setOutput(bool status) {
-		set(status);
-	}
+	inline static void setInput() {}
+	inline static void setInput(Gpio::InputType type) { inputType = type; }
+	inline static void configure(Gpio::InputType type) { inputType = type; }
+
+	inline static void setOutput() {}
+	inline static void setOutput(OutputType) {}
+	inline static void setOutput(bool status) { set(status); }
+
 	/// maps to `setInput(InputType::Floating)` or `setInput(InputType::PullUp)`
-	inline static void set() {
-		Pin::setInput(inputType);
-	}
+	inline static void set() { Pin::setInput(inputType); }
 	/// maps to `setOutput(::modm::Gpio::Low)`
-	modm_always_inline static void reset() {
-		Pin::setOutput(::modm::Gpio::Low);
-	}
-	inline static void set(bool status) {
-		if (status) { set(); }
-		else { reset(); }
-	}
-	inline static bool isSet() {
-		return (Pin::getDirection() == modm::Gpio::Direction::In);
-	}
-	modm_always_inline static modm::Gpio::Direction getDirection() {
-		return modm::Gpio::Direction::Out;
-	}
+	inline static void reset() { Pin::setOutput(::modm::Gpio::Low); }
+	inline static void set(bool status) { status ? set() : reset(); }
+	inline static bool isSet()
+	{ return (Pin::getDirection() == modm::Gpio::Direction::In); }
+
+	inline static modm::Gpio::Direction getDirection()
+	{ return modm::Gpio::Direction::Out; }
 };
 
-}	// namespace platform
-
-}	// namespace modm
-
-template< class Pin >
-modm::platform::Gpio::InputType
-modm::platform::GpioOpenDrain<Pin>::inputType(modm::platform::Gpio::InputType::Floating);
-
-#endif	// MODM_AVR_GPIO_BASE_HPP
+}	// namespace modm::platform