--- a/src/modm/platform/gpio/at90_tiny_mega/define.h
+++ b/src/modm/platform/gpio/at90_tiny_mega/define.h
@@ -11,8 +11,7 @@
  */
 // ----------------------------------------------------------------------------
 
-#ifndef	MODM_AT90_ATTINY_ATMEGA_GPIO_DEFINES_H
-#define	MODM_AT90_ATTINY_ATMEGA_GPIO_DEFINES_H
+#pragma once
 /// @cond
 
 #include <avr/io.h>
@@ -76,4 +75,4 @@
 #endif
 
 /// @endcond
-#endif	// MODM_AT90_ATTINY_ATMEGA_GPIO_DEFINES_H
+