The STM32 GPIO classes are manually generated, which create a lot of code and a lot of files. This is particularly noticable on devices with many pins, both calls to `lbuild build` and `scons` slow down significantly.

Instead I'm moving all the pin specific data into *one* long file, and using a templated `GpioStatic` class to specialize with this data at compile time. This shows significantly faster generation and compile times. The `GpioConnector` class has been adapted for the changes with more used of modern C++ TMP features. I've also reworked the GpioUnused implementation sightly as a specialization of the GpioStatic class.

I've also moved the STM32 external interrupt implementation into their own module with a new implementation based on `modm::inplace_function`. This makes a lot more sense than using EXTI via Gpios, since there are more event lines than just the GPIO ones. On AVRs the EXTI lines are coupled to the Gpios more strongly, so it stays like that.

- [x] STM32 implementation
- [x] STM32F1 implementation
- [x] AVR implementation
- [x] Merge all implementation into one template as done with GpioSet
- [x] Using more C++20 compile-time / template helpers?
- [x] Testing: on most boards that I could find.
- [x] AVR/STM32 GPIO documentation
- [x] New EXTI Implementation
- [x] EXTI documentation
- [x] More EXTI examples