@salkinium 2021-09-17
I did a simple test in each blink example:

```
rm -rf modm && time lbuild build && scons -c && time scons
```

| Blink Example  | lbuild build   | scons        |
|:---------------|:--------------:|:------------:|
| DISCO-F469     | ~5.5s -> ~1.5s | ~68s -> ~52s |
| NUCLEO-F103    | ~2.2s -> ~1.3s | ~13s -> ~12s |
| BLUE-PILL-F103 | ~1.8s -> ~1.1s | ~10s -> ~9s  |
| DISCO-F769     | ~5.3s -> ~1.5s | ~18s -> ~16s |
| NUCLEO-H743    | ~4.3s -> ~1.5s | ~19s -> ~17s |

While compile time does decrease a little, lbuild build is much much faster now. This is a good change!