@chris-durand 2021-09-17
> While compile time does decrease a little, lbuild build is much much faster now. This is a good change!

Nice! I will take a look at it later today.