@salkinium 2021-09-17
> I will take a look at it later today.

Maybe you also find a better way to encode the Gpio-peripheral-signal dependencies, dragging around the `template< Peripheral _ > Signal` thing everywhere is a bit dumb. I'll also take a look at it.