@chris-durand 2021-09-18
> Maybe you also find a better way to encode the Gpio-peripheral-signal dependencies, dragging around the `template< Peripheral _ > Signal` thing everywhere is a bit dumb. I'll also take a look at it.

I might have found a usable way to avoid the ugly `template< template< Peripheral _ > class Signal >` everywhere in the drivers.
My idea is to make the signal data plain types instead of templates and the connection data is encoded in a separate internal type `template<typename Signal, Peripheral p> struct SignalConnection` not exposed to users.

The gpio data would look like:
```
struct DataA0 {
        [...]
        // signals
	struct BitBang { using Data = DataA0; };
	struct Ch1 { using Data = DataA0; };
        [...]
	template<typename Signal, Peripheral p> struct SignalConnection;
};
[...]

// partial specialization for non-matching case, will always trigger static assert
template<Peripheral p> struct DataA0::SignalConnection<DataA0::Ch1, p> {
	static_assert(
		(p == Peripheral::Tim2),"GpioA0::Ch1 only connects to Tim2!");
};

template<> struct DataA0::SignalConnection<DataA0::Ch1, Peripheral::Tim2> {
	using Data = DataA0; static constexpr int8_t af = 2;
	static constexpr Gpio::Signal Signal = Gpio::Signal::Ch1;
};
```

I will upload a small proof-of-concept for STM32 on my fork as soon as the F1 compiles.