@chris-durand 2021-09-18
I am wondering if there is a problem with the STM32F1 alternate function configuration of this refactored version. For F1 the unspecialized version of `GpioConnector<>::connect()` does nothing and there is no partial specialization for `Peripheral::Spi2` when I generate the files for the Nucleo F103 examples. So it looks like it would not configure the alternate function mode for Spi2 when calling `connect()` from the driver.