@chris-durand 2021-09-18
@salkinium Proof-of-concept for removing the signal template template parameter: https://github.com/chris-durand/modm/commit/727fe18a94b04411c35ca8e17a402e14e0c1f2cc