@chris-durand 2021-09-19
> Yes, that's great, I'm going to cherry pick that! You're right about the F1, the connect for unspecialized peripherals must not be empty. I'll do some more testing.

It would need a one-line change in all the drivers with a connect method. I have only done it for the STM32 uart. Furthermore, the other platforms have to be adapted as well. Should I do it or will you?