@salkinium 2021-09-19
I'll work on it, also redo the AVR implementation to match.