@salkinium 2021-09-19
I've modified your prototype just a little to reduce data duplication, quite happy with this implementation. Kinda obvious in hindsight to use template specialization to code the triplet dependencies. Still very neat.
I'm on holiday until next week without hardware access, so this won't get merged until then anyways.